@@ -1,49 +1,48 @@
 {
   "dependencies": {
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-hot-loader": "4.13.0"
+    "@types/compression": "^1.7.2",
+    "@types/express": "^4.17.14",
+    "@types/node": "^18.11.9",
+    "@types/react": "^18.0.8",
+    "@types/react-dom": "^18.0.3",
+    "@typescript-eslint/eslint-plugin": "^5.59.0",
+    "@typescript-eslint/parser": "^5.59.0",
+    "@vitejs/plugin-react": "^3.0.0",
+    "compression": "^1.7.4",
+    "cross-env": "^7.0.3",
+    "eslint": "^8.38.0",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.3.4",
+    "express": "^4.18.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "sirv": "^2.0.2",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.4",
+    "vite": "^4.0.3",
+    "vite-plugin-ssr": "^0.4.133"
   },
   "devDependencies": {
-    "@babel/cli": "7.17.6",
-    "@babel/core": "7.17.9",
-    "@babel/plugin-transform-runtime": "7.17.10",
-    "@babel/preset-env": "7.16.11",
-    "@babel/preset-react": "7.16.7",
-    "@svgr/webpack": "6.2.1",
-    "autoprefixer": "10.4.4",
-    "babel-loader": "8.2.4",
-    "copy-webpack-plugin": "10.2.4",
-    "css-loader": "6.7.1",
-    "css-minimizer-webpack-plugin": "3.4.1",
-    "file-loader": "6.2.0",
-    "html-webpack-plugin": "5.5.0",
-    "image-minimizer-webpack-plugin": "3.2.3",
-    "imagemin": "8.0.1",
+    "@vheemstra/vite-plugin-imagemin": "1.0.11",
     "imagemin-gifsicle": "7.0.0",
     "imagemin-jpeg-exif-rotate": "0.0.1",
-    "imagemin-jpegoptim": "8.0.0",
+    "imagemin-mozjpeg": "10.0.0",
     "imagemin-svgo": "10.0.1",
     "imagemin-zopfli": "7.0.0",
-    "node-sass": "7.0.1",
-    "postcss": "8.4.12",
-    "postcss-line-height-px-to-unitless": "0.2.1",
-    "postcss-loader": "6.2.1",
-    "postcss-pxtorem": "6.0.0",
-    "sass-loader": "12.6.0",
-    "style-loader": "3.3.1",
-    "webpack": "5.72.0",
-    "webpack-cli": "4.9.2",
-    "webpack-dev-server": "4.8.1"
+    "sass": "1.64.1",
+    "vite-plugin-svgr": "3.2.0"
   },
   "engines": {
     "node": ">=16.14.2"
   },
-  "license": "UNLICENSED",
   "scripts": {
-    "build": "yarn && yarn clean && webpack --env production",
-    "clean": "rm -rf dist",
-    "serve": "webpack serve"
+    "build": "vite build",
+    "dev": "npm run server",
+    "lint": "eslint . --max-warnings 0",
+    "prod": "npm run lint && npm run build && npm run server:prod",
+    "server": "ts-node ./server/index.ts",
+    "server:prod": "cross-env NODE_ENV=production ts-node ./server/index.ts"
   },
   "type": "module"
 }
